feat(dump-vecs): add progress indicators
This commit adds a progress bar to the sqlite loading loop.
Useful when loading a dataset for the first time in a machine, as it
gives ETAs and throughput.

It also adds a simple elapsed time indicator when loading the Words2vec
file.

Testing: ran `dump-vecs.py` against the Google dataset.