feat(dump-hints): add progress indicators
This commit adds a progress bar to the hint generator loop.
Useful when loading a dataset for the first time in a machine, as it
gives ETAs and throughput.

It also adds a simple elapsed time indicator when loading the Words2vec
file, and some other places

Testing: ran `dump-hints.py` against the Google dataset.