feat(etl): filter hints to improve game experience
`nearby` can have a lot of similar, non-frequent words. That reduces it
usefulness as a way to find a path towards the secret word.

This commit filters the most similar words as extracted from the
Word2vec model, with a set of common spanish words. It also removes some
"boring" words based on their endings (gerunds, etc.).

Performance suffers a bit, as we now have to oversample the most similar
words, to make sure there's 1k after applying filters. Currently the
oversample factor is 100x. It might be possible to dynamically adjust
it, but that's a story for another day.

There's also a bit of cleaning.